+++ components/builder-core/src/error.rs
@@ -19,22 +19,14 @@ use std::string;
 
 use base64;
 use hab_core;
-use github_api_client;
-use git2;
-use url;
 
 #[derive(Debug)]
 pub enum Error {
     Base64Error(base64::DecodeError),
-    CannotAddCreds,
     DecryptError(String),
     EncryptError(String),
     FromUtf8Error(string::FromUtf8Error),
     HabitatCore(hab_core::Error),
-    Git(git2::Error),
-    GithubAppAuthErr(github_api_client::HubError),
-    NotHTTPSCloneUrl(url::Url),
-    UrlParseError(url::ParseError),
 }
 
 pub type Result<T> = result::Result<T, Error>;
@@ -43,20 +35,10 @@ impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let msg = match *self {
             Error::Base64Error(ref e) => format!("{}", e),
-            Error::CannotAddCreds => format!("Cannot add credentials to url"),
             Error::DecryptError(ref e) => format!("{}", e),
             Error::EncryptError(ref e) => format!("{}", e),
             Error::FromUtf8Error(ref e) => format!("{}", e),
-            Error::Git(ref e) => format!("{}", e),
-            Error::GithubAppAuthErr(ref e) => format!("{}", e),
             Error::HabitatCore(ref e) => format!("{}", e),
-            Error::NotHTTPSCloneUrl(ref e) => {
-                format!(
-                    "Attempted to clone {}. Only HTTPS clone urls are supported",
-                    e
-                )
-            }
-            Error::UrlParseError(ref e) => format!("{}", e),
         };
         write!(f, "{}", msg)
     }
@@ -66,15 +48,10 @@ impl error::Error for Error {
     fn description(&self) -> &str {
         match *self {
             Error::Base64Error(ref e) => e.description(),
-            Error::CannotAddCreds => "Cannot add credentials to url",
             Error::DecryptError(_) => "Error decrypting integration",
             Error::EncryptError(_) => "Error encrypting integration",
             Error::FromUtf8Error(ref e) => e.description(),
-            Error::Git(ref err) => err.description(),
-            Error::GithubAppAuthErr(ref err) => err.description(),
             Error::HabitatCore(ref err) => err.description(),
-            Error::NotHTTPSCloneUrl(_) => "Only HTTPS clone urls are supported",
-            Error::UrlParseError(ref err) => err.description(),
         }
     }
 }
