+++ components/builder-core/src/lib.rs
@@ -19,7 +19,6 @@ extern crate glob;
 extern crate habitat_builder_protocol as protocol;
 extern crate habitat_core as hab_core;
 extern crate habitat_net as hab_net;
-extern crate github_api_client;
 extern crate iron;
 #[macro_use]
 extern crate log;
@@ -35,8 +34,6 @@ extern crate serde;
 extern crate serde_derive;
 extern crate serde_json;
 extern crate toml;
-extern crate url;
-extern crate git2;
 
 pub mod build_config;
 pub mod data_structures;
@@ -50,7 +47,6 @@ pub mod metrics;
 pub mod package_graph;
 pub mod rdeps;
 pub mod target_graph;
-pub mod vcs;
 pub mod job;
 
 pub use error::Error;
