+++ components/builder-worker/src/lib.rs
@@ -20,12 +20,12 @@ extern crate bitflags;
 extern crate chrono;
 #[macro_use]
 extern crate features;
+extern crate git2;
 extern crate github_api_client;
 extern crate habitat_builder_protocol as protocol;
 extern crate habitat_core as hab_core;
 extern crate habitat_depot_client as depot_client;
 extern crate habitat_net as hab_net;
-extern crate git2;
 #[macro_use]
 extern crate lazy_static;
 #[macro_use]
@@ -49,6 +49,7 @@ pub mod heartbeat;
 pub mod log_forwarder;
 pub mod runner;
 pub mod server;
+pub mod vcs;
 
 pub use self::config::Config;
 pub use self::error::{Error, Result};
